backend/requirements.txt
Flask==1.1.1
Flask-HTTPAuth==3.3.0
azure-storage-blob==2.0.1

